divide performance-map implements into performance.go
